dashboard_backend/backend/requirements.txt
fastapi
uvicorn[standard]
aiohttp
websockets
pydantic
python-dotenv
